+++ lib/ui/academics/SkillsSelfEvaluationOccupationListPanel.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:neom/model/Occupation.dart';
+import 'package:neom/service/Analytics.dart';
 import 'package:neom/service/Occupations.dart';
 import 'package:neom/ui/academics/SkillsSelfEvaluationOccupationDetails.dart';
 import 'package:neom/ui/widgets/HeaderBar.dart';
@@ -186,10 +187,7 @@ class _SkillSelfEvaluationOccupationListState extends State<SkillSelfEvaluationO
                   });
                 }, icon: Styles().images.getImage('close-circle', excludeFromSemantics: true) ?? Container()),
               ),
-              IconButton(onPressed: () => setState(() {
-                _searchTerm = _searchController.text;
-                _filterOccupationList();
-              }),
+              IconButton(onPressed: _onTapSearch,
               icon: Styles().images.getImage('search', excludeFromSemantics: true) ?? Container()),
             ],
           ),
@@ -244,6 +242,15 @@ class _SkillSelfEvaluationOccupationListState extends State<SkillSelfEvaluationO
     } else {// descending
       _occupationMatchesFiltered.sort((a, b) => (b.matchPercent ?? 0).compareTo(a.matchPercent ?? 0));
     }
+  }
+
+  void _onTapSearch() {
+    Analytics().logSelect(target: "Search");
+    setState(() {
+      Analytics().logSearch(_searchController.text);
+      _searchTerm = _searchController.text;
+      _filterOccupationList();
+    });
   }
 
   void _onTapToggleSortMatchPercentage() {